+++ app/src/main/res/layout/audio_recording.xml
@@ -68,6 +68,7 @@
             android:layout_marginLeft="80dp"
             android:layout_marginTop="20dp"
             android:layout_marginBottom="10dp"
+            android:onClick="playPauseButtonClick"
             app:srcCompat="@android:drawable/ic_media_play" />
 
         <Chronometer
